Add Easy C++ Solution for 1584. Average Salary Excluding the Minimum and Maximum Salary from LeetCode

+++ LeetCode/C++/Easy/1584. Average Salary Excluding the Minimum and Maximum Salary/Solution.cpp
@@ -11,6 +11,6 @@ public:
                 min = x;
         }
 
-        return (sum - max - min)/(salary.size() - 2);
+        return double(sum - max - min)/double(salary.size() - 2);
     }
 };
